+++ poker-pals/src/components/Poker.vue
@@ -2,77 +2,14 @@
     <div class="outerFrame"> <!--the outer container of the entire page, used to define its size-->
         <UserNavbar /> <!-- the naviagation bar-->
         <div class="tableLayoutAndDisplay"> <!--a container to separate the table and inputs from the chat-->
-            <div class="tableLayout"> <!--the container holding the entire table view-->
-                <div class="buttonExitTable" v-on:click='exitTable()'> <!-- the exit table button-->
-                    <img src="../images/exit_symbol.png"/>
-                </div>
-                <div class="tableObjects"> <!--a container to hold the table and players, keeps everything centered-->
-                    <div class="tableImage"> <!--the green capsule shaped table image-->
-                        <img src="../images/table.png"/>
-                    </div>
-                    <div class="tableItems"> <!--a container to hold the pot count and community cards-->
-                        <div class="potContainer">
-                            <div class="potTag">POT</div>
-                            <div class="potAmount">{{ potTotal }}</div>
-                        </div>
-                        <div class="tableCards"> <!--the community cards container-->
-                            <div class="tableCard" id="communityCard1">
-                                <img v-bind:src="communityCards[0].src"/>
-                            </div>
-                            <div class="tableCard" id="communityCard2">
-                                <img v-bind:src="communityCards[1].src"/>
-                            </div>
-                            <div class="tableCard" id="communityCard3">
-                                <img v-bind:src="communityCards[2].src"/>
-                            </div>
-                            <div class="tableCard" id="communityCard4">
-                                <img v-bind:src="communityCards[3].src"/>
-                            </div>
-                            <div class="tableCard" id="communityCard5">
-                                <img v-bind:src="communityCards[4].src"/>
-                            </div>
-                        </div>
-                        <div class="tableDeck"> <!--image showing the deck of cards, for visuals only-->
-                            <img v-bind:src="cardFiles[0].src"/>
-                        </div>
-                    </div>
-                    <div class="players"> <!--the container holding the six players at the table, each player receives a collection of props to be displayed-->
-                        <PlayerSeat v-for="player in players" :key="player.id" v-bind:id='player.divID'
-                            v-bind:occupied="player.occupied"
-                            v-bind:classes="player.classes"
-                            v-bind:dealerStatus="player.dealerStatus" 
-                            v-bind:_id="player.id" 
-                            v-bind:cards="player.cards" 
-                            v-bind:cardReveal="cardReveal"
-                            v-bind:betValue="player.bet"
-                            v-bind:accountName="player.accountName"
-                            v-bind:accountImage="player.accountImage"
-                            v-bind:chipTotal="player.chipTotal"
-                            v-bind:action="player.action"
-                            v-bind:youTag="player.youTag"
-                            v-bind:timer="player.timer"
-                        />
-                    </div>
-                </div>
-                <div class="blindsBox">Blinds: {{ bigBlind/2 }}/{{ bigBlind }}</div> <!--shows the blinds of this particular table, TODO later bind it to a variable-->
-            </div>
-            <div class="display">
-                <div class="cheatSheetToggle" v-on:click="toggleCheatSheet()"> <!--a toggle to show or hide the cheat sheet-->
-                    <img src="../images/cheat_sheet_icon.png"/>
-                </div>
-                <div class="cheatSheet" id="cheatSheet">
-                    <img src="../images/cheatSheet.png"/>
-                </div>
-                <div class="playerCards"> <!--container holding the players personal cards-->
-                    <div class="playerCard" id="playerCard1">
-                        <img v-bind:src="myCards[0].src"/>
-                    </div>
-                    <div class="playerCard" id="playerCard2">
-                        <img v-bind:src="myCards[1].src"/>
-                    </div>
-                </div>
-                <PlayerInputs v-bind:bigBlind="bigBlind"/>
-            </div>
+            <TableLayout 
+                v-bind:potTotal="potTotal"
+                v-bind:communityCards="communityCards"
+                v-bind:players="players"
+                v-bind:cardReveal="cardReveal"
+                v-bind:bigBlind="bigBlind"
+            />
+            <Display v-bind:myCards="myCards" v-bind:bigBlind="bigBlind"/>
         </div>
         <Chat v-bind:full="chatFull"/> <!--the chat container holding all items related to messaging other players-->
         <ReportPocket v-if="!chatFull"
@@ -97,166 +34,6 @@
         width: 70%;
         height: 93%;
     }
-    .tableLayout{
-        position: relative;
-        width: 100%;
-        height: 70%;
-        background: #b7b7b7;
-    }
-    .buttonExitTable{
-        position: absolute;
-        width: 4%;
-        height: 10%;
-        margin: 1% 0% 0% 1%;
-        background-color: transparent;
-        border: none;
-        cursor: pointer;
-    }
-    .buttonExitTable:active {
-        transform: translateY(4%);
-    }
-    .buttonExitTable img{
-        float: left;
-    }
-    .tableObjects{
-        position: absolute;
-        width: 100%;
-        height: 79%;
-        margin: 5.3% 0% 0% 0%;
-    }
-    .tableImage{
-        position: absolute;
-        width: 78%;
-        height: 75%;
-        margin: 5.75% 0% 0% 11.5%;
-        background: transparent;
-    }
-    .tableItems{
-        position: absolute;
-        width: 50%;
-        height: 15%;
-        background: transparent;
-        margin: 17.75% 0% 0% 25%;
-    }
-    .potContainer{
-        position: absolute;
-        width: 25%;
-        height: 100%;
-        border-radius: 15%/30%;
-        background: #bf214b;
-        border: 1px solid black;
-        margin: -0.25% 0% 0% -0.125%;
-    }
-    .potTag{
-        color: white;
-        font-size: 80%;
-        margin: 0%;
-        font-size: 1.25vw;
-    }
-    .potAmount{
-        color: white;
-        font-size: 1vw;
-    }                   
-    .tableCards{
-        position: absolute;
-        width: 61%;
-        height: 100%;
-        background: transparent;
-        margin: 0% 0% 0% 25.5%;
-    }
-    .tableCard{
-        float: left;
-        width: 17.5%;
-        height: 100%;
-        margin: 0% 0% 0% 2%;
-        border-radius: 15%;
-    }
-    .tableDeck{
-        position: absolute;
-        width: 11.5%;
-        height: 100%;
-        background: transparent;
-        margin: 0% 0% 0% 88.5%;
-        border-radius: 15%;
-    }
-    .players{
-        position: absolute;
-        width: 60%;
-        height: 110%;
-        margin: -2% 0% 0% 20%;
-    }
-    #playerSeat1{
-        margin: 0%;
-    }
-    #playerSeat2{
-        margin: 0% 0% 0% 12.5%;
-    }
-    #playerSeat3{
-        margin: 0% 0% 0% 12.5%;
-    }
-    #playerSeat4{
-        position: absolute;
-        margin: 54% 0% 0% 75%;
-    }
-    #playerSeat5{
-        position: absolute;
-        margin: 54% 0% 0% 37.5%;
-    }
-    #playerSeat6{
-        position: absolute;
-        margin: 54% 0% 0% 0%;
-    }
-    .blindsBox{
-        position: absolute;
-        width: 20%;
-        height: 6%;
-        background: transparent;
-        margin: 49.2% 0% 0% 80%;
-        font-size: 1vw;
-        text-align: center;
-        font-weight: 1000;
-    }
-    .display{
-        width: 100%;
-        height: 30%;
-        background: #eeeeee;
-        margin: 0%;
-    }
-    .cheatSheetToggle{
-        float: left;
-        width: 3%;
-        height: 14%;
-        margin: 18.25% 0% 0% 1%;
-        background-color: transparent;
-        cursor: pointer;
-    }
-    .cheatSheetToggle:active {
-        transform: translateY(4%);
-    }
-    .cheatSheetToggle img{
-        float: left;
-        width: 100%;
-        height: 100%;
-    }
-    .cheatSheet{
-        float: left;
-        width: 30%;
-        height: 90%;
-        background: transparent;
-        margin: 1% 0% 0% 1%;
-    }
-    .playerCards{
-        float: left;
-        width: 28%;
-        height: 90%;
-        background: transparent;
-        margin: 1% 0% 0% 1%;
-    }
-    .playerCard{
-        float: left;
-        width: 45%;
-        margin: 5.5% 0% 0% 3%;
-    }
     
     img{
         width: 100%;
@@ -266,11 +43,11 @@
 
 <script>
 
-import PlayerSeat from './pokerComponents/PlayerSeat.vue';
-import UserNavbar from './UserNavbar.vue';
+import UserNavbar from './navbars/UserNavbar.vue';
 import Chat from './pokerComponents/Chat.vue';
 import ReportPocket from './pokerComponents/ReportPocket.vue';
-import PlayerInputs from './pokerComponents/PlayerInputs.vue';
+import TableLayout from './pokerComponents/TableLayout.vue';
+import Display from './pokerComponents/Display.vue';
 
 
 import io from "socket.io-client";
@@ -278,11 +55,11 @@ import io from "socket.io-client";
 export default {
     name: "Poker",
     components: {
-        PlayerSeat,
         UserNavbar,
         Chat,
         ReportPocket,
-        PlayerInputs,
+        TableLayout,
+        Display,
     },
     created() {
         this.socket = io("http://localhost:3000");
@@ -422,7 +199,7 @@ export default {
         
         this.socket.on('showdown', msgJSON => { // the showdown has begun, now all player cards are being shown
             let msg = JSON.parse(msgJSON);
-            for(let i = 0; i < 6; i++){
+            for(let i = 0; i < this.players.length; i++){
                 this.players[i].cards = {
                     card1: {src: this.cardFiles.find(file => file.name === msg[i][0]).src}, 
                     card2: {src: this.cardFiles.find(file => file.name === msg[i][1]).src}
@@ -437,7 +214,7 @@ export default {
 
         this.socket.on('reset', () => { // after a round a new game will begin shortly, reset the table to a state that is ready for a new round
             // each player will reset the ui for a new game
-            for(let i = 0; i < 6; i++){
+            for(let i = 0; i < this.players.length; i++){
                 this.players[i].cards = {card1: {src: this.cardFiles[0].src}, card2: {src: this.cardFiles[0].src}};
             }
             for(let i = 0; i < 5; i++){
@@ -458,10 +235,10 @@ export default {
         // later the server will send the true cards at the show down
         this.myCards = [{src: this.cardFiles[1].src}, {src: this.cardFiles[1].src}];
         
-        for(let i = 0; i < 6; i++){
+        for(let i = 0; i < this.players.length; i++){
             this.players[i].cards = {card1: {src: this.cardFiles[0].src}, card2: {src: this.cardFiles[0].src}};
         }
-        for(let i = 0; i < 5; i++){
+        for(let i = 0; i < this.communityCards.length; i++){
             this.communityCards[i].src = this.cardFiles[1].src;
         }
     },
@@ -498,15 +275,6 @@ export default {
         },
         exitTable(){ // allows a player to leave the table, player will forfeit any chips in the pot
             // TODO: when all pages are linked this function will be implemented
-        },
-        toggleCheatSheet(){
-            if(this.cheatSheetOpen){
-                document.getElementById('cheatSheet').style.opacity = 0.0;
-            }
-            else{
-                document.getElementById('cheatSheet').style.opacity = 1.0;
-            }
-            this.cheatSheetOpen = !this.cheatSheetOpen;
         },
         setRaiseToValue(value){
             this.raiseToValue = value;
