+++ poker-pals/src/components/pokerComponents/Display.vue
@@ -1,0 +1,107 @@
+<template>
+    <div class="display">
+        <div class="cheatSheetToggle" v-on:click="toggleCheatSheet()"> <!--a toggle to show or hide the cheat sheet-->
+            <img src="../../images/cheat_sheet_icon.png"/>
+        </div>
+        <div class="cheatSheet" id="cheatSheet">
+            <img src="../../images/cheatSheet.png"/>
+        </div>
+        <div class="playerCards"> <!--container holding the players personal cards-->
+            <div class="playerCard" id="playerCard1">
+                <img v-bind:src="myCards[0].src"/>
+            </div>
+            <div class="playerCard" id="playerCard2">
+                <img v-bind:src="myCards[1].src"/>
+            </div>
+        </div>
+        <PlayerInputs v-bind:bigBlind="bigBlind"/>
+    </div>
+</template>
+
+<style scoped>
+    .display{
+        width: 100%;
+        height: 30%;
+        background: #eeeeee;
+        margin: 0%;
+    }
+    .cheatSheetToggle{
+        float: left;
+        width: 3%;
+        height: 14%;
+        margin: 18.25% 0% 0% 1%;
+        background-color: transparent;
+        cursor: pointer;
+    }
+    .cheatSheetToggle:active {
+        transform: translateY(4%);
+    }
+    .cheatSheetToggle img{
+        float: left;
+        width: 100%;
+        height: 100%;
+    }
+    .cheatSheet{
+        float: left;
+        width: 30%;
+        height: 90%;
+        background: transparent;
+        margin: 1% 0% 0% 1%;
+    }
+    .playerCards{
+        float: left;
+        width: 28%;
+        height: 90%;
+        background: transparent;
+        margin: 1% 0% 0% 1%;
+    }
+    .playerCard{
+        float: left;
+        width: 45%;
+        margin: 5.5% 0% 0% 3%;
+    }
+    
+    img{
+        width: 100%;
+        height: 100%;
+    }
+</style>
+
+<script>
+import PlayerInputs from './PlayerInputs.vue';
+
+export default {
+    name: "Display",
+    components: {
+        PlayerInputs,
+    },
+    props: [
+        'myCards', 'bigBlind',
+    ],
+    data() {
+        return {
+            cheatSheetOpen: true, // toggle to diplay cheat sheet or not
+        }
+    },
+    methods:{
+        toggleCheatSheet(){
+            if(this.cheatSheetOpen){
+                document.getElementById('cheatSheet').style.opacity = 0.0;
+            }
+            else{
+                document.getElementById('cheatSheet').style.opacity = 1.0;
+            }
+            this.cheatSheetOpen = !this.cheatSheetOpen;
+        },
+        makeDecision(action){
+            this.$parent.makeDecision(action);
+        },
+        toggleCheckFoldButton(checkFold){
+            this.$parent.toggleCheckFoldButton(checkFold);
+        },
+        setRaiseToValue(value){
+            this.$parent.setRaiseToValue(value);
+        }
+    }
+};
+</script>
