+++ poker-pals/src/components/pokerComponents/PlayerInputs.vue
@@ -3,16 +3,7 @@
         <button class="inputButton1 centerText" id="buttonFold" v-on:click="$parent.makeDecision('FOLD')">FOLD</button>
         <button class="inputButton1 centerText" id="buttonCheck" v-on:click="$parent.makeDecision('CHECK')">CHECK</button>
         <button class="inputButton1 centerText" id="buttonCall" v-on:click="$parent.makeDecision('CALL')">CALL</button>
-        <div class="raiseToggle"> <!--the container holding the raise input, it contains two buttons, a slider, and a value display-->
-            <div class="raiseButtonOuter" id="buttonMinus" v-on:click="decrementRaise()">
-                <div class="raiseButton centerText">-</div>
-            </div>
-            <div class="raiseValue" id="raiseValue">2000</div>
-            <input class="raiseScroll" id="slider" type="range" min="2000" max="100000" value="2000"> <!--TODO: set these using props-->
-            <div class="raiseButtonOuter" id="buttonPlus" v-on:click="incrementRaise()">
-                <div class="raiseButton centerText">+</div>
-            </div>
-        </div>
+        <RaiseToggle v-bind:bigBlind="bigBlind"/>
         <button class="inputButton1 centerText" id="buttonRaise" v-on:click="$parent.makeDecision('RAISE')">RAISE</button>
         <button class="inputButton2 centerText" id="buttonCheckFold" v-on:click="toggleCheckFoldButton()">CHECK/FOLD</button>
         <button class="inputButton1 centerText" id="buttonAllIn" v-on:click="$parent.makeDecision('ALL IN')">ALL IN</button>
@@ -20,19 +11,7 @@
 </template>
 
 <style scoped>
-
-    /* Chrome, Safari, Edge, Opera */
-    input::-webkit-outer-spin-button,
-    input::-webkit-inner-spin-button {
-    -webkit-appearance: none;
-    margin: 0;
-    }
-
-    /* Firefox */
-    input[type=number] {
-    -moz-appearance: textfield;
-    }
-
+    
     .playerInputs{
         float: left;
         width: 32%;
@@ -67,57 +46,6 @@
     #buttonCall{
         margin: 0% 0% 0% 5%;
     }
-    .raiseToggle{
-        float: left;
-        width: 65%;
-        height: 25%;
-        margin: 6.75% 0% 0% 0%;
-    }
-    .raiseButtonOuter{
-        float: left;
-        width: 20%;
-        height: 100%;
-        background: black;
-        cursor: pointer;
-    }
-    .raiseButtonOuter:active{
-        transform: translateX(4%);
-    }
-    .raiseButton{
-        width: 95%;
-        height: 95%;
-        background: white;
-        font-weight: 700;
-        font-size: 1vw;
-        margin: 2.5% 0% 0% 2.5%;
-    }
-    #buttonMinus{
-        margin: 0%;
-    }
-    .raiseValue{
-        position: absolute;
-        width: 12.25%;
-        height: 3%;
-        margin: 0% 0% 0% 4.25%;
-        font-weight: 700;
-        font-size: 1vw;
-    }
-    .raiseScroll{
-        float: left;
-        width: 60%;
-        height: 40%;
-        margin: 13% 0% 0% 0%;
-        background: transparent;
-    }
-    .raiseScroll::-moz-range-thumb{
-        width: 25%;
-        height: 100%;
-        background: #eeeeee;
-        cursor: pointer;
-    }
-    #buttonPlus{
-        margin: 0%;
-    }
     #buttonRaise{
         margin: 6.75% 0% 0% 5%;
     }
@@ -142,9 +70,13 @@
 </style>
 
 <script>
+import RaiseToggle from './RaiseToggle.vue';
 
 export default {
     name: "PlayerInputs",
+    components: {
+        RaiseToggle,
+    },
     props: [
         'bigBlind',
     ],
@@ -153,12 +85,6 @@ export default {
             // TODO: bind data with props
             checkFold: false,
         }
-    },
-    mounted(){
-        // TODO:
-    },
-    watch: {
-        // TODO:
     },
     methods:{
         toggleCheckFoldButton(){ // toggle the check/fold button, this will allow the client to automatically send an action without user input
@@ -172,35 +98,8 @@ export default {
             }
             this.$parent.toggleCheckFoldButton(this.checkFold);
         },
-        incrementRaise(){
-            // cap this value if it goes over what the player has left
-            let slider = document.getElementById("slider");
-            let output = document.getElementById("raiseValue");
-            let value = parseInt(slider.value);
-            // add up to the big blind to even out the number
-            slider.value = (value + (this.bigBlind - (value % this.bigBlind))).toString();
-            output.innerHTML = slider.value;
-            this.sendRaiseToValue();
-        },
-        decrementRaise(){
-            // cap this value if it goes over what the player has left
-            let slider = document.getElementById("slider");
-            let output = document.getElementById("raiseValue");
-            let value = parseInt(slider.value);
-            // subtract up to the big blind
-            let remove = value % this.bigBlind;
-            if(remove === 0){
-                remove = this.bigBlind;
-            }
-            slider.value = (value - remove).toString();
-            output.innerHTML = slider.value;
-            this.sendRaiseToValue();
-        },
-        sendRaiseToValue(){
-            let slider = document.getElementById("slider");
-            let value = parseInt(slider.value);
+        setRaiseToValue(value){
             this.$parent.setRaiseToValue(value);
-            return value;
         },
     }
 };
