+++ server/PokerTableObjects/HelperFunctions.js
@@ -1,0 +1,5 @@
+module.exports = {
+    calculate: function(){
+        
+    }
+};
